Refactored code to reduce code duplicates
