+++ content/_code-samples/build-a-wallet/js/3_account.js
@@ -1,34 +1,12 @@
 const { app, BrowserWindow, ipcMain} = require('electron')
 const path = require('path')
 const xrpl = require("xrpl")
+const { prepareReserve, prepareAccountData } = require('./library/3_helpers')
 
 const TESTNET_URL = "wss://s.altnet.rippletest.net:51233"
 
 let reserveBaseXrp = null, reserveIncrementXrp = null
 
-const prepareLedgerData = (ledger) => {
-    reserveBaseXrp = xrpl.dropsToXrp(ledger.reserve_base)
-    reserveIncrementXrp = xrpl.dropsToXrp(ledger.reserve_inc)
-
-    return ledger
-}
-
-const prepareAccountData = (rawAccountData) => {
-    const numOwners = rawAccountData.OwnerCount || 0
-
-    let xrpReserve = null
-    if (reserveBaseXrp && reserveIncrementXrp) {
-        //TODO: Decimal?
-        xrpReserve = reserveBaseXrp + (reserveIncrementXrp * numOwners)
-    }
-
-    return {
-        classicAddress: rawAccountData.Account,
-        xAddress: xrpl.classicAddressToXAddress(rawAccountData.Account, false, true),
-        xrpBalance: xrpl.dropsToXrp(rawAccountData.Balance),
-        xrpReserve: xrpReserve
-    }
-}
 const createWindow = () => {
 
     const appWindow = new BrowserWindow({
@@ -49,6 +27,8 @@ const main = async () => {
 
     ipcMain.on('address-entered', async (event, address) =>  {
 
+        let reserve = null
+
         const client = new xrpl.Client(TESTNET_URL)
 
         await client.connect()
@@ -61,8 +41,8 @@ const main = async () => {
 
         //
         client.on("ledgerClosed", async (ledger) => {
-            const ledgerData = prepareLedgerData(ledger)
-            appWindow.webContents.send('update-ledger-data', ledgerData)
+            reserve = prepareReserve(ledger)
+            appWindow.webContents.send('update-ledger-data', ledger)
         })
 
         // Wait for transaction on subscribed account and re-request account data
@@ -73,7 +53,7 @@ const main = async () => {
                 "ledger_index": transaction.ledger_index
             }
             const accountInfoResponse = await client.request(accountInfoRequest)
-            const accountData = prepareAccountData(accountInfoResponse.result.account_data)
+            const accountData = prepareAccountData(accountInfoResponse.result.account_data, reserve)
             appWindow.webContents.send('update-account-data', accountData)
         })
